# swagger/dev.json
{
  "swagger": "2.0",
  "info": {
    "title": "Distributed Big Data service Open Platform",
    "version": "1.0.0",
    "description": "Distributed Big Data service Open Platform back-end API",
    "contact": {
      "name": "Yimin Lin",
      "email": "linyimin520812@gmail.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/",
  "paths": {
    "/databases": {
      "get": {
        "description": "Get All mysql databases",
        "tags": [
          ["databases"]
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/databases/tables": {
      "get": {
        "tags": [
          ["databases"]
        ],
        "description": "Get all tables' name from a database",
        "produces": ["application/json"],
        "parameters": [{
          "name": "databaseName",
          "description": "database name to database for geting all tables.",
          "in": "query",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/databases/tables/{name}": {
      "get": {
        "tags": [
          ["databases"]
        ],
        "description": "Get all tables' fields from a table",
        "produces": ["application/json"],
        "parameters": [{
          "name": "name",
          "description": "Table name to table for geting all fields.",
          "in": "path",
          "required": true,
          "type": "string"
        }, {
          "name": "databaseName",
          "description": "database name to database for geting all tables.",
          "in": "query",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/service/parameter-type": {
      "get": {
        "tags": [
          ["service"]
        ],
        "description": "Get parameters' type and responses' type",
        "produces": ["application/json"],
        "parameters": [{
          "name": "sql",
          "description": "A sql staetment.",
          "in": "query",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/service/generate": {
      "post": {
        "tags": [
          ["service"]
        ],
        "description": "Get parameters' type and responses' type",
        "produces": ["application/json"],
        "parameters": [{
          "name": "method",
          "description": "The request method of the service which will be generated.",
          "in": "query",
          "required": true,
          "type": "string"
        }, {
          "name": "path",
          "description": "The URL of the service.",
          "in": "query",
          "required": true,
          "type": "string"
        }, {
          "name": "name",
          "description": "The name of the service.",
          "in": "query",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/sql/generate": {
      "post": {
        "tags": [
          ["sql"]
        ],
        "description": "Generate a sql statement",
        "produces": ["application/json"],
        "parameters": [{
          "name": "select",
          "in": "body",
          "description": "Target field want to get",
          "required": true,
          "schema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }, {
          "name": "from",
          "in": "body",
          "description": "database and table names",
          "required": true,
          "schema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }, {
          "name": "where",
          "in": "body",
          "description": "Condition of sql",
          "required": true,
          "schema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }, {
          "name": "groupBy",
          "in": "body",
          "description": "Group by condition",
          "required": false,
          "schema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }, {
          "name": "orderBy",
          "in": "body",
          "description": "Order by condition",
          "required": false,
          "schema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }, {
          "name": "limit",
          "in": "body",
          "description": "limit",
          "required": false,
          "schema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/sql/test": {
      "get": {
        "tags": [
          ["sql"]
        ],
        "description": "Test generated sql",
        "produces": ["application/json"],
        "parameters": [{
          "name": "sql",
          "description": "A sql staetment.",
          "in": "query",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {},
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}